requirements.txt
torch>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
PyQt5>=5.15.0
pyqtgraph>=0.13.0
numpy>=1.24.0
PyYAML>=6.0
